pyproject.toml
[project]
name = "btwin"
version = "0.1.0"
description = "B-TWIN: AI partner that remembers your thoughts"
requires-python = ">=3.11"
dependencies = [
    "litellm>=1.55",
    "chromadb>=0.6",
    "mcp[cli]>=1.2",
    "typer>=0.15",
    "pyyaml>=6.0",
    "pydantic>=2.0",
    "rich>=13.0",
]

[project.scripts]
btwin = "btwin.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/btwin"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
